refactor: rename modal types import to BaseModal.types for consistency
