+++ src/presentation/widgets/modals/BaseModal.types.ts
